init SA
