+++ src/py/flwr/common/secure_aggregation/crypto/symmetric_encryption.py
@@ -1,0 +1,94 @@
+# Copyright 2020 Adap GmbH. All Rights Reserved.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+# ==============================================================================
+"""Symmetric encryption."""
+
+
+import base64
+from typing import Tuple
+
+from cryptography.fernet import Fernet
+from cryptography.hazmat.primitives import hashes, serialization
+from cryptography.hazmat.primitives.asymmetric import ec
+from cryptography.hazmat.primitives.kdf.hkdf import HKDF
+
+
+# Generate private and public key pairs with Cryptography
+def generate_key_pairs() -> (
+    Tuple[ec.EllipticCurvePrivateKey, ec.EllipticCurvePublicKey]
+):
+    sk = ec.generate_private_key(ec.SECP384R1())
+    pk = sk.public_key()
+    return sk, pk
+
+
+# Serialize private key
+def private_key_to_bytes(sk: ec.EllipticCurvePrivateKey) -> bytes:
+    return sk.private_bytes(
+        encoding=serialization.Encoding.PEM,
+        format=serialization.PrivateFormat.PKCS8,
+        encryption_algorithm=serialization.NoEncryption(),
+    )
+
+
+# Deserialize private key
+def bytes_to_private_key(b: bytes) -> ec.EllipticCurvePrivateKey:
+    return serialization.load_pem_private_key(data=b, password=None)
+
+
+# Serialize public key
+def public_key_to_bytes(pk: ec.EllipticCurvePublicKey) -> bytes:
+    return pk.public_bytes(
+        encoding=serialization.Encoding.PEM,
+        format=serialization.PublicFormat.SubjectPublicKeyInfo,
+    )
+
+
+# Deserialize public key
+def bytes_to_public_key(b: bytes) -> ec.EllipticCurvePublicKey:
+    return serialization.load_pem_public_key(data=b)
+
+
+# Generate shared key by exchange function and key derivation function
+# Key derivation function is needed to obtain final shared key of exactly 32 bytes
+def generate_shared_key(
+    sk: ec.EllipticCurvePrivateKey, pk: ec.EllipticCurvePublicKey
+) -> bytes:
+    # Generate a 32 byte urlsafe(for fernet) shared key
+    # from own private key and another public key
+    sharedk = sk.exchange(ec.ECDH(), pk)
+    derivedk = HKDF(
+        algorithm=hashes.SHA256(),
+        length=32,
+        salt=None,
+        info=None,
+    ).derive(sharedk)
+    return base64.urlsafe_b64encode(derivedk)
+
+
+# Authenticated Encryption =========================================================
+
+
+# Encrypt plaintext with Fernet. Key must be 32 bytes.
+def encrypt(key: bytes, plaintext: bytes) -> bytes:
+    # key must be url safe
+    f = Fernet(key)
+    return f.encrypt(plaintext)
+
+
+# Decrypt ciphertext with Fernet. Key must be 32 bytes.
+def decrypt(key: bytes, token: bytes):
+    # key must be url safe
+    f = Fernet(key)
+    return f.decrypt(token)
