--- src/py/flwr/common/secure_aggregation/message_handler.py
@@ -1,8 +1,0 @@
-from abc import ABC
-
-from flwr.common.typing import SecureAggregation
-
-
-class SecureAggregationHandler(ABC):
-    def handle_secure_aggregation(self, sa: SecureAggregation) -> SecureAggregation:
-        ...
